integrations/requirements.txt
flask==3.0.1
coverage==6.5.0
setuptools~=70.0.0
wheel~=0.38.4
pip==23.3.1
boto3==1.17.54
Django==4.2.17
mssql-django==1.5
pyodbc==5.1.0
pytz==2022.6
pytest-django==4.5.2
requests==2.32.3
requests_mock==1.9.1
attrs==19.2.0
pytest==7.2.0
pytest-cov==4.0.0
git+https://github.com/part-os/core-python.git@7e8d9ed26c22c804225116350fe33fc2763578cb#egg=core-python
sentry-sdk==2.8.0
safety==2.3.5
cryptography==42.0.6
flake8==3.9.1
pendulum==3.0.0
pyyaml==6.0.1
apscheduler==3.7.0
pycountry==20.7.3
pgeocode==0.3.0
google-api-python-client==2.11.0
fuzzywuzzy==0.18.0
simplejson==3.17.5
pytest-mock==3.10.0
debugpy==1.5.1
pydantic==1.10.13
numpy==1.23.5
hubspot-api-client==4.0.6
xmltodict==0.13.0
openpyxl==3.0.9
pandas==1.4.2
xlsxwriter==3.2.0
xlutils==2.0.0
xlrd==2.0.1
